Created some UI Structuring components
The game now launches into a debug menu when built in debug mode,
instead of the main menu
Wrapped the exit application code button in an #if desktop
Added a first draft of the Container and Container Layout data
structure. They allow quick UIs to be constructed and laid out similar
to how swing components were done in Java. (Except that this is optional
and can have an entire game around it.)