Switched some graphics around
Made a new level file for the new background that we were given.
Swapped Khonjin from a red rectangle, to his placeholder sprite, and
scaled him down a bit because he's big as fuck
There's a known issue where Khonjin can walk off the side of the level.
I have an idea to fix it, but I don't have the time right now.